+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### v 0.0.6 (1/20/14)
 
-- added restart method
+- added resume method
 
 ### v 0.0.5 (1/17/14)